feature(CalendarView): simplifying month view
